+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## Release 5.0.M14
 
+* #613 Exception on include when entity id is unreadable
 * #615 Swagger: ignore UriInfo if @Parameter(hidden=true) is set
 * #616 Swagger: duplicate param properties
 * #617 Proper handling of ObjectId annotated with @AgId